🐛Fix: type과 data 분리
